+++ FMS/Pages/Maintenance/AddComplianceOfficer.cshtml
@@ -1,7 +1,6 @@
 ﻿@page
-@using FMS.Domain.Entities.Users
-@using FMS.Pages.Admin
-@model FMS.Pages.Admin.AddComplianceOfficerModel
+@using FMS.Pages.Maintenance
+@model FMS.Pages.Maintenance.AddComplianceOfficerModel
 @section Scripts {
     <partial name="_ValidationScriptsPartial" />
 }
