+++ FMS/Pages/Maintenance/AddFacilityType.cshtml
@@ -1,27 +1,18 @@
-﻿@page "{id:Guid?}"
-@using FMS.Pages.Admin
-@model FMS.Pages.Admin.EditFacilityTypeModel
+﻿@page
+@using FMS.Pages.Maintenance
+@model FMS.Pages.Maintenance.AddFacilityTypeModel
 
 @section Scripts {
     <partial name="_ValidationScriptsPartial" />
 }
 
 @{
-    ViewData["Title"] = $"Edit {MaintenanceOptions.FacilityType}: {Model.FacilityType.Name}";
+    ViewData["Title"] = $"Add {MaintenanceOptions.FacilityType}";
 }
-<h1>
-    @ViewData["Title"]
-    @if (!Model.FacilityType.Active)
-    {
-        <span class="text-muted">(Removed)</span>
-    }
-</h1>
-
+<h1>@ViewData["Title"]</h1>
 <div class="container bg-light py-3">
     <form method="post">
         <div asp-validation-summary="ModelOnly" class="alert alert-danger" role="alert"></div>
-        <input type="hidden" asp-for="Id" />
-        <input type="hidden" asp-for="FacilityType.Active" />
         <div class="form-row">
             <div class="form-group col-md-3">
                 <label asp-for="FacilityType.Name">Code <span class="text-danger">*</span></label>
@@ -29,14 +20,14 @@
                 <span asp-validation-for="FacilityType.Name" class="text-danger"></span>
             </div>
             <div class="form-group col-md">
-                <label asp-for="FacilityType.Description">Description</label>
+                <label asp-for="FacilityType.Description"></label>
                 <input asp-for="FacilityType.Description" class="form-control" />
                 <span asp-validation-for="FacilityType.Description" class="text-danger"></span>
             </div>
         </div>
         <p class="text-danger">* denotes a required field</p>
         <hr />
-        <button type="submit" class="btn btn-primary col-sm-3 mb-3 mb-sm-0">Save Changes</button>
+        <button id="SubmitButton" type="submit" class="btn btn-primary col-md-4 mb-3 mb-md-0">Add New @MaintenanceOptions.FacilityType</button>
         <a asp-page="Index" class="btn btn-outline-secondary col-md-2" asp-route-handler="select"
            asp-route-MaintenanceSelection="@MaintenanceOptions.FacilityType">
             Cancel
