+++ FMS/Pages/Maintenance/AddOrganizationalUnit.cshtml
@@ -1,0 +1,31 @@
+﻿@page
+@using FMS.Pages.Maintenance
+@model FMS.Pages.Maintenance.AddOrganizationalUnitModel
+
+@section Scripts {
+    <partial name="_ValidationScriptsPartial" />
+}
+
+@{
+    ViewData["Title"] = $"Add {MaintenanceOptions.OrganizationalUnit}";
+}
+<h1>@ViewData["Title"]</h1>
+<div class="container bg-light py-3">
+    <form method="post">
+        <div asp-validation-summary="ModelOnly" class="alert alert-danger" role="alert"></div>
+        <div class="form-row">
+            <div class="form-group col-md-10">
+                <label asp-for="OrganizationalUnit.Name">Organizational Unit Name <span class="text-danger">*</span></label>
+                <input asp-for="OrganizationalUnit.Name" class="form-control" />
+                <span asp-validation-for="OrganizationalUnit.Name" class="text-danger"></span>
+            </div>
+        </div>
+        <p class="text-danger">* denotes a required field</p>
+        <hr />
+        <button id="SubmitButton" type="submit" class="btn btn-primary col-md-5 mb-3 mb-md-0">Add New Organizational Unit</button>
+        <a asp-page="Index" class="btn btn-outline-secondary col-md-2" asp-route-handler="select"
+           asp-route-MaintenanceSelection="@MaintenanceOptions.OrganizationalUnit">
+            Cancel
+        </a>
+    </form>
+</div>
