+++ FMS/Pages/Maintenance/DeleteBudgetCode.cshtml
@@ -1,6 +1,6 @@
 ﻿@page "{id:Guid?}"
-@using FMS.Pages.Admin
-@model FMS.Pages.Admin.DeleteBudgetCodeModel
+@using FMS.Pages.Maintenance
+@model FMS.Pages.Maintenance.DeleteBudgetCodeModel
 @{
     ViewData["Title"] = $"Remove Budget Code: {Model.Code}";
 }
