+++ FMS/Pages/Maintenance/DeleteOrganizationalUnit.cshtml
@@ -1,6 +1,6 @@
 ﻿@page "{id:Guid?}"
-@using FMS.Pages.Admin
-@model FMS.Pages.Admin.DeleteOrganizationalUnitModel
+@using FMS.Pages.Maintenance
+@model FMS.Pages.Maintenance.DeleteOrganizationalUnitModel
 @{
     ViewData["Title"] = $"Remove Organizational Unit: {Model.Name}";
 }
