+++ FMS/Pages/Maintenance/DetailsComplianceOfficer.cshtml
@@ -1,6 +1,6 @@
 ﻿@page "{id:Guid?}"
-@using FMS.Pages.Admin
-@model FMS.Pages.Admin.DetailsComplianceOfficerModel
+@using FMS.Pages.Maintenance
+@model FMS.Pages.Maintenance.DetailsComplianceOfficerModel
 @{
     ViewData["Title"] = $"Compliance Officer: {Model.DisplayName}";
 }
