+++ FMS/Pages/Maintenance/EditFacilityType.cshtml
@@ -1,18 +1,27 @@
-﻿@page
-@using FMS.Pages.Admin
-@model FMS.Pages.Admin.AddFacilityTypeModel
+﻿@page "{id:Guid?}"
+@using FMS.Pages.Maintenance
+@model FMS.Pages.Maintenance.EditFacilityTypeModel
 
 @section Scripts {
     <partial name="_ValidationScriptsPartial" />
 }
 
 @{
-    ViewData["Title"] = $"Add {MaintenanceOptions.FacilityType}";
+    ViewData["Title"] = $"Edit {MaintenanceOptions.FacilityType}: {Model.FacilityType.Name}";
 }
-<h1>@ViewData["Title"]</h1>
+<h1>
+    @ViewData["Title"]
+    @if (!Model.FacilityType.Active)
+    {
+        <span class="text-muted">(Removed)</span>
+    }
+</h1>
+
 <div class="container bg-light py-3">
     <form method="post">
         <div asp-validation-summary="ModelOnly" class="alert alert-danger" role="alert"></div>
+        <input type="hidden" asp-for="Id" />
+        <input type="hidden" asp-for="FacilityType.Active" />
         <div class="form-row">
             <div class="form-group col-md-3">
                 <label asp-for="FacilityType.Name">Code <span class="text-danger">*</span></label>
@@ -20,14 +29,14 @@
                 <span asp-validation-for="FacilityType.Name" class="text-danger"></span>
             </div>
             <div class="form-group col-md">
-                <label asp-for="FacilityType.Description"></label>
+                <label asp-for="FacilityType.Description">Description</label>
                 <input asp-for="FacilityType.Description" class="form-control" />
                 <span asp-validation-for="FacilityType.Description" class="text-danger"></span>
             </div>
         </div>
         <p class="text-danger">* denotes a required field</p>
         <hr />
-        <button id="SubmitButton" type="submit" class="btn btn-primary col-md-4 mb-3 mb-md-0">Add New @MaintenanceOptions.FacilityType</button>
+        <button type="submit" class="btn btn-primary col-sm-3 mb-3 mb-sm-0">Save Changes</button>
         <a asp-page="Index" class="btn btn-outline-secondary col-md-2" asp-route-handler="select"
            asp-route-MaintenanceSelection="@MaintenanceOptions.FacilityType">
             Cancel
