+++ FMS/Pages/Maintenance/Index.cshtml
@@ -1,6 +1,6 @@
 ﻿@page
-@using FMS.Pages.Admin
-@model FMS.Pages.Admin.IndexModel
+@using FMS.Pages.Maintenance
+@model FMS.Pages.Maintenance.IndexModel
 @{
     if (string.IsNullOrEmpty(Model.MaintenanceSelection))
     {
