+++ FMS/Pages/Shared/_MenuPartial.cshtml
@@ -31,7 +31,7 @@
                         <a class="dropdown-item" asp-page="/Users/Index">FMS Users</a>
                         @if (User.IsInRole(UserRoles.SiteMaintenance))
                         {
-                            <a class="dropdown-item" asp-page="/Admin/Index">Site Maintenance</a>
+                            <a class="dropdown-item" asp-page="/Maintenance/Index">Site Maintenance</a>
                         }
                     </div>
                 </li>
